Moved industry data out of EveData monolith
